% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcfwa_geometry.R
\name{bcfwa_geometry}
\alias{bcfwa_geometry}
\title{BCFWA Stream Geometry}
\usage{
bcfwa_geometry(
  bcfwa = NA,
  upstream = 701794363,
  downstream = 701773410,
  epsg = 26910
)
}
\arguments{
\item{bcfwa}{sf dataframe. Imported BCFWA streamline geometry. Import
directly as gdb or gpkg to avoid losing the z (or M) elevation geometry.
Do not convert to shapefile. Usually this object is imported by a BCFWA
group code. See example.}

\item{upstream}{numeric. Upstream linear
feature ID from BCFWA. See field (LINEAR_FEATURE_ID)... or spatial point}

\item{downstream}{numeric. Downstream linear feature ID
from BCFWA. See field (LINEAR_FEATURE_ID)... or spatial point}

\item{epsg}{Numeric. EPSG code for local UTM projection system (see:
https://spatialreference.org/ref/epsg/ for details).}
}
\value{
An sf dataframe object of the target creek/river and a dataframe
of coordinates.
}
\description{
Extract geometry from BCFWA data for target river
}
\details{
Extracts x,y and z geometry from the BCFWA (British Columbia
Freshwater Atlas) for a target river/creek of interest. Specify upstream
and downstream LFID (linear feature ID).
See FWA_STREAM_NETWORKS_SP.gdb at the FTP link.
(FTP: ftp://ftp.geobc.gov.bc.ca/sections/outgoing/bmgs/FWA_Public).
Additional steps involve reordering line features (where necessary).
Processing is done for a single
}
\examples{
\dontrun{
library(streamgis)
# Import BCFWA streamlines from FWA_STREAM_NETWORKS_SP.gdb
# Find target group code (layer) for area of interest (e.g., 'LNIC')
# center_line <- st_read("./path/to/BCFWA/FWA_STREAM_NETWORKS_SP.gdb",
# layer = "LNIC")

# or continue with default provided for tutorial
fname <- system.file("extdata", "bcfwa.gpkg", package="streamgis")
bcfwa <- sf::st_read(fname)

ds <- bcfwa_geometry(bcfwa = bcfwa,
upstream = 701794363,
downstream = 701773410,
epsg = 26910)

# ds_path: is the original geometry
# coordinates: is the x,y,z coordinates
names(ds)

# View the GIS data
plot(sf::st_geometry(ds$ds_path))

# View coordinates for longitudial profile
df <- ds$coordinates
plot(df$us_distance_m/1000, df$Z, type = 'l',
xlab = "Upstream Distance (km)", ylab = "Elevation (m)")

# Beginning of line
head(df, 3)

# End of line
tail(df, 3)

# Validate length
seg <- bcfwa[bcfwa$GNIS_NAME == "Coldwater River", ]
tot_length <- sf::st_length(seg)
as.numeric(sum(tot_length))



}

}
