% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{clip_line_between_points}
\alias{clip_line_between_points}
\title{Helper: Clip line between two points}
\usage{
clip_line_between_points(line, pt1_coords, pt2_coords)
}
\description{
Helper: Clip line between two points
}
\keyword{internal}
