% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{clip_line_from_start}
\alias{clip_line_from_start}
\title{Helper: Clip line from snapped point onwards (removes start)}
\usage{
clip_line_from_start(line, point_coords)
}
\description{
Helper: Clip line from snapped point onwards (removes start)
}
\keyword{internal}
