% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{clip_line_to_end}
\alias{clip_line_to_end}
\title{Helper: Clip line up to snapped point (removes end)}
\usage{
clip_line_to_end(line, point_coords)
}
\description{
Helper: Clip line up to snapped point (removes end)
}
\keyword{internal}
