% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{create_stream_network}
\alias{create_stream_network}
\title{Helper: Create sfnetwork from streamlines}
\usage{
create_stream_network(streamlines)
}
\description{
Helper: Create sfnetwork from streamlines
}
\keyword{internal}
