% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{extract_path_segments}
\alias{extract_path_segments}
\title{Helper: Extract and combine path segments from streamlines_split}
\usage{
extract_path_segments(streamlines_split, snapped_pt1, snapped_pt2, path_result)
}
\description{
Helper: Extract and combine path segments from streamlines_split
}
\keyword{internal}
