% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_lines_at_points.R
\name{find_position_on_line}
\alias{find_position_on_line}
\title{Helper: Find position along line (0-1 scale)}
\usage{
find_position_on_line(line_coords, point_coords)
}
\description{
Helper: Find position along line (0-1 scale)
}
\keyword{internal}
