% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{find_shortest_path_network}
\alias{find_shortest_path_network}
\title{Helper: Find shortest path between two points on network
Returns edge indices that map to row numbers in streamlines}
\usage{
find_shortest_path_network(network, streamlines, snapped_pt1, snapped_pt2)
}
\description{
Helper: Find shortest path between two points on network
Returns edge indices that map to row numbers in streamlines
}
\keyword{internal}
