% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utm_epsg.R
\name{get_utm_epsg}
\alias{get_utm_epsg}
\title{Get UTM epsg}
\usage{
get_utm_epsg(sf_object)
}
\arguments{
\item{sf_object}{sf dataframe.}
}
\value{
Numeric epsg code.
}
\description{
Gets the local UTM epsg code
}
\details{
This function takes a sf dataframe object and returns the local UTM.
If single polygon or line epsg code will be sampled from the centroid. If
sf_object consists of multiple objects in the form the bounding box
}
\examples{
\dontrun{

library(streamgis)
# Import a simple stream center line
# center_line <- st_read("./path/to/my/file.gpkg", layer = "layer name")
fname <- system.file("extdata", "center_line.gpkg", package="streamgis")
center_line <- sf::st_read(fname)
get_utm_epsg(sf_object = center_line)
}

}
