% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data_to_cross_sections.R
\name{join_data_to_cross_sections}
\alias{join_data_to_cross_sections}
\title{Join External Data to Cross Sections}
\usage{
join_data_to_cross_sections(
  to_join = NA,
  type = "closest_points",
  cross_section_lines = NA,
  points_on_line = NA,
  epsg = 26910
)
}
\arguments{
\item{to_join}{sf dataframe. Extrnal data to join to cross sections. Spatial
data of class sf with geometry type of point, line or polygon. Must be
projected to same CRS as \code{cross_section_lines} and \code{points_on_line}. Joined
datat will be returned as a dataframe to \code{cbind} to the original cross shore
profiles or lines.}

\item{type}{character string. Either "closest_points" or
"perpendicular_lines". "closest_points" is designed to join point data to the
closest cross section. "perpendicular_lines" is designed to join streamline
data linked to \code{to_join} line geometry objects running parallel to stream and
perpendicular to cross sections. "perpendicular_lines" will be joined with an
intersection and "closest_points" will be joined by distance.}

\item{cross_section_lines}{sf dataframe. Object returned from
\code{cross_section_lines}.}

\item{points_on_line}{Optional. sf dataframe. Ideally bbject returned from
\code{points_on_line}. If NA \code{st_centroid()} will be used to sample a single
point along each \code{cross_section_lines}.}

\item{epsg}{Numeric. EPSG code for local UTM projection system (see:
https://spatialreference.org/ref/epsg/ for details).}
}
\value{
An dataframe of the linked data from \code{to_join} linked to each row in
the parent \code{points_on_line} or \code{cross_section_lines} dataframe.
}
\description{
Join External Data to Cross Sectional Lines
}
\details{
Join external data to cross sectional lines. External data is
provided as a \code{to_join} sf datamframe object (e.g., points, lines or
polygons). Join\code{type} is specified and then function returns a dataframe of
the joined data to the parent dataframe of the \code{cross_section_lines}. Return
type is a dataframe to avoid column name issues.
}
\examples{
\dontrun{
}

}
