% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_on_line.R
\name{points_on_line}
\alias{points_on_line}
\title{Generate Points on Line}
\usage{
points_on_line(center_line = NA, point_spacing = 50, epsg = 26910)
}
\arguments{
\item{center_line}{sf dataframe. Stream center line spatial data. Must be
imported as an sf datamfrace object of LINESTRING or MULTILINESTRING. The
object must have a column named \code{id} to represent the id field. The object
must also be projected to a local UTM projection with units of meters.}

\item{point_spacing}{Numeric. Point spacing distance in units of meters.}

\item{epsg}{Numeric. EPSG code for local UTM projection system (see:
https://spatialreference.org/ref/epsg/ for details).}
}
\value{
An sf dataframe object of points sampled along stream center lines.
}
\description{
Samples points on stream centerline
}
\details{
Samples points at a regular interval along stream center line
these points can be used to construct stream cross sectional profiles
and provide utility for other applications in hydrology and aquatic
habitat assessments. Output fields include line segement id (l_id),
point id (p_id, unique to each line segement). Distance (for each point
on each line segement from line segment origin) and group according to
whether point represents are \code{start} point, \code{end} point or \code{sample} point.
Filter out start and end points if they are not needed.
}
\examples{
\dontrun{

library(streamgis)
# Import a simple stream center line
# center_line <- st_read("./path/to/my/file.gpkg", layer = "layer name")
# or use default provided for tutorial
fname <- system.file("extdata", "center_line.gpkg", package="streamgis")
center_line <- sf::st_read(fname)
plot(sf::st_geometry(center_line))

# Sample points along line
pol <- suppressWarnings({ points_on_line(center_line,
  point_spacing = 100, epsg = 26910) })

# Plot to visualize
plot(sf::st_geometry(center_line[center_line$id == 1, ]))
plot(sf::st_geometry(pol[pol$l_id == 1, ]), add = TRUE, col = "red")


}

}
