% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_at_confluences.R
\name{snap_endpoints_to_lines}
\alias{snap_endpoints_to_lines}
\title{Snap Line Endpoints to Nearby Lines}
\usage{
snap_endpoints_to_lines(streamlines, snap_tolerance)
}
\arguments{
\item{streamlines}{An sf LINESTRING object.}

\item{snap_tolerance}{Maximum distance for snapping (in CRS units).}
}
\value{
A list with:
\describe{
\item{lines}{The streamlines with snapped endpoints}
\item{n_snapped}{Number of endpoints that were snapped}
}
}
\description{
Snaps line endpoints that are close to (but don't touch) other
lines to the nearest point on those lines. This fixes common digitization
errors where tributaries overshoot or undershoot the mainstem.
}
\keyword{internal}
