% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{snap_point_to_streamline}
\alias{snap_point_to_streamline}
\title{Helper: Snap a single point to nearest streamline}
\usage{
snap_point_to_streamline(streamlines, point)
}
\description{
Helper: Snap a single point to nearest streamline
}
\keyword{internal}
