% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{split_line_at_point}
\alias{split_line_at_point}
\title{Helper: Split a single line at one point}
\usage{
split_line_at_point(line, point_coords)
}
\description{
Helper: Split a single line at one point
}
\keyword{internal}
