% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{split_line_at_two_points}
\alias{split_line_at_two_points}
\title{Helper: Split a single line at two points}
\usage{
split_line_at_two_points(line, pt1_coords, pt2_coords)
}
\description{
Helper: Split a single line at two points
}
\keyword{internal}
