% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_path.R
\name{split_streamlines_at_points}
\alias{split_streamlines_at_points}
\title{Helper: Split streamlines at snapped point locations}
\usage{
split_streamlines_at_points(streamlines_cast, snapped_pt1, snapped_pt2)
}
\description{
Helper: Split streamlines at snapped point locations
}
\keyword{internal}
