% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_at_confluences.R
\name{unique_points_within_tolerance}
\alias{unique_points_within_tolerance}
\title{Remove Duplicate Points Within Tolerance}
\usage{
unique_points_within_tolerance(points_sfc, tolerance)
}
\description{
Uses spatial clustering to efficiently remove duplicate points.
Points within tolerance distance are merged by taking the centroid of each cluster.
}
\keyword{internal}
